Testing Update
Re-tooled testing scripts to use parametrized test data, and conduct more robust testing. Bumped version to 0.1.7
