+++ setup.py
@@ -54,6 +54,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/geozeke/parser201',
-    version='0.1.6',
+    version='0.1.7',
     zip_safe=False,
 )
